# swaggerJson/swagger-ledger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/classify/{id}": {
      "delete": {
        "operationId": "ClassifyController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "数据id",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer JWT Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "path": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "删除",
        "tags": [
          "账本-分类"
        ]
      },
      "put": {
        "operationId": "ClassifyController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "数据id",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer JWT Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClassifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "path": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          },
          {
            "Authorization": []
          }
        ],
        "summary": "更新",
        "tags": [
          "账本-分类"
        ]
      },
      "get": {
        "operationId": "ClassifyController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "数据id",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer JWT Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "path": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "$ref": "#/components/schemas/DetailClassifyRes"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          },
          {
            "Authorization": []
          }
        ],
        "summary": "详情",
        "tags": [
          "账本-分类"
        ]
      }
    },
    "/classify": {
      "post": {
        "operationId": "ClassifyController_add",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer JWT Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "path": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CreateClassifyRes"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          },
          {
            "Authorization": []
          }
        ],
        "summary": "新增",
        "tags": [
          "账本-分类"
        ]
      },
      "get": {
        "operationId": "ClassifyController_list",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer JWT Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "页码，从 1 开始",
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "default": 1,
              "example": 1
            }
          },
          {
            "description": "每页数量，默认为 10",
            "name": "size",
            "in": "query",
            "required": true,
            "schema": {
              "default": 10,
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "path": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ListClassifyRes"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          },
          {
            "Authorization": []
          }
        ],
        "summary": "分页列表",
        "tags": [
          "账本-分类"
        ]
      }
    },
...
    "/ledger": {
      "post": {
        "operationId": "LedgerController_add",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer JWT Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLedgerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "path": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CreateLedgerRes"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "新增",
        "tags": [
          "账本-账本"
        ]
      },
      "get": {
        "operationId": "LedgerController_list",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer JWT Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "页码，从 1 开始",
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "default": 1,
              "example": 1
            }
          },
          {
            "description": "每页数量，默认为 10",
            "name": "size",
            "in": "query",
            "required": true,
            "schema": {
              "default": 10,
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "path": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ListLedgerRes"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          },
          {
            "Authorization": []
          }
        ],
        "summary": "分页列表",
        "tags": [
          "账本-账本"
        ]
      }
    },
...
  },
  "info": {
    "title": "ledger Example",
    "description": "ledger 接口文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "请输入 JWT Token（示例：eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.xxx）"
      }
    },
    "schemas": {
...
      "ClassifyRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "应用id",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "应用名称",
            "example": "应用名称"
          },
          "userId": {
            "type": "number",
            "description": "创建人id",
            "example": 1
          },
          "ledgerId": {
            "type": "number",
            "description": "账本id",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2025/12/27 20:59:01"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2025/12/27 20:59:01"
          }
        },
        "required": [
          "id",
          "name",
          "userId",
          "ledgerId",
          "createdAt",
          "updatedAt"
        ]
      },
      "LedgerRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "id",
            "example": 1
          },
          "userId": {
            "type": "number",
            "description": "创建人id",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "账单名称",
            "example": "账单名称"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2025/12/27 20:59:01"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2025/12/27 20:59:01"
          }
        },
        "required": [
          "id",
          "userId",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateClassifyRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "应用id",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "应用名称",
            "example": "应用名称"
          },
          "userId": {
            "type": "number",
            "description": "创建人id",
            "example": 1
          },
          "ledgerId": {
            "type": "number",
            "description": "账本id",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2025/12/27 20:59:01"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2025/12/27 20:59:01"
          }
        },
        "required": [
          "id",
          "name",
          "userId",
          "ledgerId",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateClassifyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "分类名称",
            "example": "分类名称"
          },
          "ledgerId": {
            "type": "number",
            "description": "账本id",
            "example": 1
          }
        },
        "required": [
          "name",
          "ledgerId"
        ]
      },
      "UpdateClassifyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "分类名称",
            "example": "分类名称"
          }
        },
        "required": [
          "name"
        ]
      },
      "DetailClassifyRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "应用id",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "应用名称",
            "example": "应用名称"
          },
          "userId": {
            "type": "number",
            "description": "创建人id",
            "example": 1
          },
          "ledgerId": {
            "type": "number",
            "description": "账本id",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2025/12/27 20:59:01"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2025/12/27 20:59:01"
          }
        },
        "required": [
          "id",
          "name",
          "userId",
          "ledgerId",
          "createdAt",
          "updatedAt"
        ]
      },
      "ListClassifyRes": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "页码"
          },
          "size": {
            "type": "number",
            "description": "每页数量"
          },
          "total": {
            "type": "number",
            "description": "数据总数"
          },
          "list": {
            "description": "分类信息",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassifyRes"
            }
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "list"
        ]
      },
      "CreateLedgerRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "id",
            "example": 1
          },
          "userId": {
            "type": "number",
            "description": "创建人id",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "账单名称",
            "example": "账单名称"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2025/12/27 20:59:01"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2025/12/27 20:59:01"
          }
        },
        "required": [
          "id",
          "userId",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateLedgerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "账本名称",
            "example": "账本名称"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "ListLedgerRes": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "页码"
          },
          "size": {
            "type": "number",
            "description": "每页数量"
          },
          "total": {
            "type": "number",
            "description": "数据总数"
          },
          "list": {
            "description": "账本信息",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerRes"
            }
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "list"
        ]
      }
    }
  }
}